Move lattice boundary tests to dispatch lane
