+++ tests/math/lattices/test_lattice_operations.py
@@ -5,9 +5,7 @@ from __future__ import annotations
 import pytest
 from pydantic import ValidationError
 
-from jacobian.catalog.catalog import Catalog
 from jacobian.catalog.models import OperationDomainValidationError
-from jacobian.dispatch import OperationRequestValidationError, invoke_operation
 from jacobian.math.lattices._hnf import compute_hermite_normal_form
 from jacobian.math.lattices._lattice import reduce_lattice_basis
 from jacobian.math.lattices._models import (
@@ -374,33 +372,6 @@ def test_hermite_request_rejects_order_33_identity() -> None:
     assert exc_info.value.errors()[0]["type"] == "lattice.budget_exceeded"
 
 
-def test_lattice_reduction_admits_before_lll_backend() -> None:
-    """A 33-by-33 identity must not reach FLINT or fail inside the result model."""
-
-    basis = IntegerMatrix.model_validate(
-        {"entries": _identity_entries(MAX_MATRIX_DIMENSION + 1)}
-    )
-    request = LatticeReductionRequest.model_construct(basis=basis)
-    with pytest.raises(OperationDomainValidationError) as exc_info:
-        reduce_lattice_basis(request)
-    error = exc_info.value.errors()[0]
-    assert error["type"] == "lattice.budget_exceeded"
-    assert error["loc"] == ("basis",)
-    assert str(MAX_MATRIX_DIMENSION) in error["msg"]
-
-
-def test_hermite_admits_before_hnf_backend() -> None:
-    matrix = IntegerMatrix.model_validate(
-        {"entries": _identity_entries(MAX_MATRIX_DIMENSION + 1)}
-    )
-    request = HermiteNormalFormRequest.model_construct(matrix=matrix)
-    with pytest.raises(OperationDomainValidationError) as exc_info:
-        compute_hermite_normal_form(request)
-    error = exc_info.value.errors()[0]
-    assert error["type"] == "lattice.budget_exceeded"
-    assert error["loc"] == ("matrix",)
-
-
 def test_lattice_reduction_accepts_order_32_identity() -> None:
     request = LatticeReductionRequest.model_validate(
         {"basis": {"entries": _identity_entries(MAX_MATRIX_DIMENSION)}}
@@ -419,18 +390,6 @@ def test_hermite_accepts_order_32_identity() -> None:
     result = compute_hermite_normal_form(request)
     assert result.normal_form.entries == tuple(
         tuple(row) for row in _identity_entries(MAX_MATRIX_DIMENSION)
-    )
-
-
-def test_dispatch_rejects_lll_above_the_lattice_axis() -> None:
-    with pytest.raises(OperationRequestValidationError) as exc_info:
-        invoke_operation(
-            "lattice.basis.reduce",
-            {"basis": {"entries": _identity_entries(MAX_MATRIX_DIMENSION + 1)}},
-            Catalog.open(),
-        )
-    assert f"limited to {MAX_MATRIX_DIMENSION} rows and columns" in str(
-        exc_info.value.cause
     )
